fix: extract _ensure_reporter_participant to behaviors layer (ISSUE-1421)
Move _ensure_reporter_participant and _upgrade_participant_to_accepted from
vultron.core.use_cases.received.case._helpers into
vultron.core.behaviors.case.nodes.participant.common, eliminating the
BT→use_cases import-direction violation (AGENTS.md BT-IDM-02).

EnsureReporterParticipantAtAcceptedNode._bootstrap previously used a
deferred import inside update() as a workaround — now uses a clean
module-level import from the behaviors layer.

Backward-compat re-exports remain in _helpers.py (# noqa: F401) and
received/case/__init__.py so existing callers are unaffected.

Follow-up: #1428 tracks a pre-existing sibling violation in common.py
(_as_id / _report_phase_status_id still imported from use_cases._helpers).
